In progress

+++ common/src/main/java/ru/yofik/athena/common/domain/model/notification/DeleteMessageNotification.kt
@@ -1,5 +1,5 @@
 package ru.yofik.athena.common.domain.model.notification
 
 data class DeleteMessageNotification(
-    val ids: List<Int>
+    val ids: List<Long>
 ) : Notification(NotificationType.DELETED_MESSAGES)
